+++ src/mid_level/mid_level_layer.py
